refactor(define-router): compress the util packaging
Fold config types, serialization, and build metadata into one module,
move path2regexp next to the other path helpers, and inline the tiny
task runner into the build handler so the split earns less ceremony.
