--- packages/waku/src/router/define-router-utils/build-metadata.ts
@@ -1,5 +1,0 @@
-export const DEFINE_ROUTER_METADATA = {
-  serializableConfigs: 'defineRouter:serializableConfigs',
-  cachedElements: 'defineRouter:cachedElements',
-  path2moduleIds: 'defineRouter:path2moduleIds',
-} as const;
