+++ packages/waku/src/router/isomorphic-utils/path-spec.ts
@@ -170,3 +170,20 @@ export const getPathMapping = (
   }
   return mapping;
 };
+
+const escapeRegExp = (s: string) => s.replace(/[\\^$.*+?()[\]{}|]/g, '\\$&');
+
+export const path2regexp = (path: PathSpec) => {
+  const parts = path.map((item) => {
+    if (item.type === 'literal') {
+      return escapeRegExp(item.name);
+    } else if (item.type === 'group') {
+      const prefix = escapeRegExp(item.prefix ?? '');
+      const suffix = escapeRegExp(item.suffix ?? '');
+      return `${prefix}([^/]+)${suffix}`;
+    } else {
+      return `(.*)`;
+    }
+  });
+  return `^/${parts.join('/')}$`;
+};
